Add support for host names in the address field
This adds a new DNS resolution function. Users can now enter v4 addresses or hostnames. V6 addresses aren't supported yet, but could be still indirectly used if the input is a hostname which resolves to a AAAA record, for example.

This also improves the alert box and error handling around connection problems.

The saved default will be whatever the user inputs into the text box. If it's a hostname, we will save that. Name resolution is always done on any hostname.